+++ redisvl/extensions/session_manager/standard_session.py
@@ -3,28 +3,13 @@ from typing import Any, Dict, List, Optional, Union
 from redis import Redis
 
 from redisvl.extensions.session_manager import BaseSessionManager
-from redisvl.extensions.session_manager.schema import ChatMessage
+from redisvl.extensions.session_manager.schema import (
+    ChatMessage,
+    StandardSessionIndexSchema,
+)
 from redisvl.index import SearchIndex
 from redisvl.query import FilterQuery
 from redisvl.query.filter import Tag
-from redisvl.schema.schema import IndexSchema
-
-
-class StandardSessionIndexSchema(IndexSchema):
-
-    @classmethod
-    def from_params(cls, name: str, prefix: str):
-
-        return cls(
-            index={"name": name, "prefix": prefix},  # type: ignore
-            fields=[  # type: ignore
-                {"name": "role", "type": "tag"},
-                {"name": "content", "type": "text"},
-                {"name": "tool_call_id", "type": "tag"},
-                {"name": "timestamp", "type": "numeric"},
-                {"name": "session_tag", "type": "tag"},
-            ],
-        )
 
 
 class StandardSessionManager(BaseSessionManager):
